model

+++ pkg/framework/class/sortedset/item.go
@@ -2,5 +2,5 @@ package sortedset
 
 type Item interface {
 	Key() string
-	Val() int64
+	Val() int
 }
